Dummy GUI
GUI mockup in a few minutes. It does not do much but it shows how
with a few lines of code a future contributor can write complete
the GUI. Raising custom exceptions from the backend (actions) and
catching them in the GUI would probably the best idea to share the
code and failure messages between both is probably a good idea.