use even more uncommon DISPLAY number
sometimes the X-server gets confused or something and thinks that
it is running on :9 but it isn't (tested, running only on :0.0)
so use something even more uncommon - :13. basicly, even if you
have X running on one of your TTYs (keyboard shorcuts from F1-F12)
it is unlikely (rare) that you are running 13 X-servers. worst
case scenario random.random() can be used to generate random
number from actions.xnest() and then pass it to misc.system_chroot()
as xnest argument.