various improvements, mostly xnest related
1. lib.misc assumed that when whereis() is called the first argument
is relative path not absolute, convert the passed argument to
basename to ensure that checking will not fail in such cases

2. split the command in lib.misc.chroot_exec() as it may consist of
program and arguments, something that subprocess.check_call() is not
able to handle properly (unless the first argument is full path and
it still must be list or constant). this was detected as a failure
while trying to run gnome-session on official Ubuntu 14.04 ISO
where the xsession Exec command is "gnome-session --session=ubuntu"

3. make sure that Xephyr is terminated if an error acures in the
chroot procedure or command, again detected while bumping on 2.