stop divertign /sbin/initctl in chroot, fixes issue #63
as bad as this change can be, if /sbin/init is diverted upgrading
packages that rely on initctl fail (e.g. systemd-services). On the
other hand if initctl is not diverted host services may be affected
from the chroot (upstart uses D-Bus, and access to it is granted
for the nested X-session). Upstart has a container feature but I'm
afraid to tell it that it is in container as it uses file in /run
and /var/run (symlink to /run) is bind mounted for the nested
X-session too. In any case, do not blame me when your host starts
to behave strange after upgrading chroot.