do not polute the current environment on chroot
setting env. variables in the current process env. instead of the
subprocess can lead to various issues, including not being able to
edit the sources.list because after xnest the DESKTOP variable is
set to :13 which is no longer valid. this change also seems to fix
an issue where on Xubuntu the XDG menu did not display all menu
entries, now this is no longer the case (all applications are shown)