use unicode() instead of str() to normalize strings (issue #93)
unicode has always been a problem both with Python 2 and 3. with
PyQt strings returned by dialogs and such are QString which have
to be normalized to be able to write them into the config and
all. Altough this fixes some of the issues it can not do more
than handle strings properly _internally_, ConfigPraser still
fails because it attempts to convert the key to string via str()
thus any chanages to the config will _not_ apply but atleast the
actions will execute now (after an error message) when paths have
unicode characters in them. feel free to improve on this.