Altered the way the rebuild script handled kernels and checksums.
Working from my past local improvements to ubuntu-builder, I've ported them to customizer, as I prefer python.
https://github.com/kamilion/kamikazi-deploy/blob/master/resources/buildscripts/Build.patched-for-xen

Fixes in this commit:
If EFI is determined to be in use, vmlinuz will be hardlinked to vmlinuz.efi, saving a few megs on the ISO.
If found, a Xen kernel will be added to the media in the correct place. Same for ipxe and memtest86+ soon?
As a fix for ubuntu-mini-remix's integrity check, a SHA256SUMS file will be regenerated if found.
This should shore up UMR in respect to #76, outside of the memtest86+ issue, which we can fix.
Compression parameters are now separated from mksquashfs, and a default blocksize of ~1MB is applied.
Editing this can be added to the GUI later, for now, xz's branch-call-jump filters were enabled for x86 code.
In addition, Many comments were added to clarify sections of the ISO rebuild code.