This has been an annoyance for a while.
Customizer will, by default, use the release name and version as the
iso's filename. Editing /etc/lsb-release to change the filename ends
up to break other things, including apt. So, fix that by providing
an override in the form of /etc/livecd-release, which, if exists,
will override the contents of /etc/lsb-release when it's read.